+++ src/test/resources/org/csanchez/jenkins/plugins/kubernetes/pipeline/declarativeFromYaml.groovy
@@ -1,11 +1,12 @@
 pipeline {
   agent {
     kubernetes {
-      label 'mypod'
+      label 'declarativefromyaml-pod'
       yaml """
 metadata:
   labels:
     some-label: some-label-value
+    class: KubernetesDeclarativeAgentTest
 spec:
   containers:
   - name: jnlp