+++ src/test/resources/org/csanchez/jenkins/plugins/kubernetes/pipeline/declarativeFromYamlFile.groovy
@@ -1,0 +1,28 @@
+pipeline {
+  agent {
+    kubernetes {
+      label 'declarativefromyamlfile-pod'
+      yamlFile 'declarativeYamlFile.yml'
+    }
+  }
+  options {
+    // Because there's no way for the container to actually get at the git repo on the disk of the box we're running on.
+    skipDefaultCheckout(true)
+  }
+  stages {
+    stage('Run maven') {
+      steps {
+        sh 'set'
+        sh "echo OUTSIDE_CONTAINER_ENV_VAR = ${CONTAINER_ENV_VAR}"
+        container('maven') {
+          sh 'echo MAVEN_CONTAINER_ENV_VAR = ${CONTAINER_ENV_VAR}'
+          sh 'mvn -version'
+        }
+        container('busybox') {
+          sh 'echo BUSYBOX_CONTAINER_ENV_VAR = ${CONTAINER_ENV_VAR}'
+          sh '/bin/busybox'
+        }
+      }
+    }
+  }
+}